Restore cosim process
This commit introduces slight modifications to the `tcp_consumer.cpp`
design unit which allow the cosim process to sucessfully complete.

Notice that by reverting the changes to the TCP session introduced in
previous commit the cosim hangs again. This means that the cosim hangs
when an inbound TCP packet triggers a collection (i.e session == 0x40)
or if the `end_of_packet` branch is enabled in `tcp_consumer.cpp`.